+++ rust/candid_parser/tests/assets/ok/float.d.ts
@@ -1,0 +1,11 @@
+import type { Principal } from '@icp-sdk/core/principal';
+import type { ActorMethod } from '@icp-sdk/core/agent';
+import type { IDL } from '@icp-sdk/core/candid';
+
+export interface _SERVICE {
+  'identity32' : ActorMethod<[number], number>,
+  'to_f32' : ActorMethod<[number], number>,
+  'to_f64' : ActorMethod<[number], number>,
+}
+export declare const idlFactory: IDL.InterfaceFactory;
+export declare const init: (args: { IDL: typeof IDL }) => IDL.Type[];